backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn
PyPDF2==3.0.1
python-docx==1.1.0
